pyproject.toml
[project]
name = "watchtower-backend"
version = "0.1.0"
description = "WATCHTOWER backend"
readme = "README.md"
authors = [
    { name = "arukanus", email = "arukanus@mail.ru" }
]
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "anthropic>=0.86.0",
    "elevenlabs>=2.39.0",
    "fastapi>=0.135.1",
    "greenlet>=3.3.2",
    "httpx>=0.28.1",
    "langchain-anthropic>=1.4.0",
    "langgraph>=1.1.3",
    "luffa-bot-python-sdk>=0.1.2",
    "numpy>=2.4.3",
    "orjson>=3.11.7",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.42.0",
]

[project.scripts]
watchtower-backend = "watchtower_backend:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.7",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "B",
    "UP",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
